refactor(api): move read models constants campaign types to campaign constants
