+++ api/db/database-builder/factory/build-assessment-campaign-for-skills.js
@@ -1,6 +1,6 @@
 import { buildCampaign } from './build-campaign.js';
 import { buildCampaignSkill } from './build-campaign-skill.js';
-import { CampaignTypes } from '../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignTypes } from '../../../src/prescription/shared/domain/constants.js';
 
 const buildAssessmentCampaignForSkills = function (attributes, skillSet) {
   attributes.type = CampaignTypes.ASSESSMENT;
