+++ api/db/database-builder/factory/build-assessment-campaign.js
@@ -1,5 +1,5 @@
 import { buildAssessmentCampaignForSkills } from './build-assessment-campaign-for-skills.js';
-import { CampaignTypes } from '../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignTypes } from '../../../src/prescription/shared/domain/constants.js';
 
 const buildAssessmentCampaign = function (attributes, skills = [{ id: 'skill', name: 'skillName' }]) {
   attributes.type = CampaignTypes.ASSESSMENT;
