+++ api/lib/domain/models/index.js
@@ -24,9 +24,8 @@ import { CampaignLearningContent } from './CampaignLearningContent.js';
 import { CampaignParticipant } from './CampaignParticipant.js';
 import { CampaignParticipation } from './CampaignParticipation.js';
 import { CampaignParticipationResult } from './CampaignParticipationResult.js';
-import { CampaignParticipationStatuses } from '../../../src/prescription/shared/domain/constants.js';
+import { CampaignParticipationStatuses, CampaignTypes } from '../../../src/prescription/shared/domain/constants.js';
 import { CampaignToStartParticipation } from './CampaignToStartParticipation.js';
-import { CampaignTypes } from '../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
 import { CertifiableBadgeAcquisition } from './CertifiableBadgeAcquisition.js';
 import { CertifiableProfileForLearningContent } from './CertifiableProfileForLearningContent.js';
 import { CertificationAnswerStatusChangeAttempt } from './CertificationAnswerStatusChangeAttempt.js';
