+++ api/lib/domain/read-models/CampaignToJoin.js
@@ -1,5 +1,5 @@
 import { Assessment } from '../../../src/shared/domain/models/Assessment.js';
-import { CampaignTypes } from '../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignTypes } from '../../../src/prescription/shared/domain/constants.js';
 
 class CampaignToJoin {
   constructor({
