+++ api/lib/domain/read-models/OrganizationLearnerActivity.js
@@ -1,4 +1,4 @@
-import { CampaignTypes } from '../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignTypes } from '../../../src/prescription/shared/domain/constants.js';
 import lodash from 'lodash';
 
 const { countBy } = lodash;
