+++ api/lib/infrastructure/repositories/campaign-participation-overview-repository.js
@@ -1,7 +1,6 @@
 import { knex } from '../../../db/knex-database-connection.js';
-import { CampaignTypes } from '../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
 import { fetchPage } from '../utils/knex-utils.js';
-import { CampaignParticipationStatuses } from '../../../src/prescription/shared/domain/constants.js';
+import { CampaignParticipationStatuses, CampaignTypes } from '../../../src/prescription/shared/domain/constants.js';
 import { CampaignParticipationOverview } from '../../domain/read-models/CampaignParticipationOverview.js';
 
 const findByUserIdWithFilters = async function ({ userId, states, page }) {
