+++ api/lib/infrastructure/repositories/campaign-participation-repository.js
@@ -1,5 +1,4 @@
-import { CampaignParticipationStatuses } from '../../../src/prescription/shared/domain/constants.js';
-import { CampaignTypes } from '../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignParticipationStatuses, CampaignTypes } from '../../../src/prescription/shared/domain/constants.js';
 import { knex } from '../../../db/knex-database-connection.js';
 import * as knowledgeElementRepository from './knowledge-element-repository.js';
 import * as knowledgeElementSnapshotRepository from './knowledge-element-snapshot-repository.js';
