+++ api/scripts/prod/create-assessment-campaigns-for-sco.js
@@ -1,7 +1,7 @@
 import bluebird from 'bluebird';
 import _ from 'lodash';
 import { knex, disconnect } from '../../db/knex-database-connection.js';
-import { CampaignTypes } from '../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignTypes } from '../../src/prescription/shared/domain/constants.js';
 import * as codeGenerator from '../../lib/domain/services/code-generator.js';
 import * as campaignUpdateValidator from '../../src/prescription/campaign/domain/validators/campaign-update-validator.js';
 import * as campaignRepository from '../../src/prescription/campaign/infrastructure/repositories/campaign-administration-repository.js';
