+++ api/scripts/prod/create-profiles-collection-campaigns.js
@@ -1,6 +1,6 @@
 import bluebird from 'bluebird';
 import { knex, disconnect } from '../../db/knex-database-connection.js';
-import { CampaignTypes } from '../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignTypes } from '../../src/prescription/shared/domain/constants.js';
 import * as codeGenerator from '../../lib/domain/services/code-generator.js';
 import * as campaignUpdateValidator from '../../src/prescription/campaign/domain/validators/campaign-update-validator.js';
 import * as campaignRepository from '../../src/prescription/campaign/infrastructure/repositories/campaign-administration-repository.js';
