--- api/src/prescription/campaign/domain/read-models/CampaignTypes.js
@@ -1,6 +1,0 @@
-const CampaignTypes = {
-  ASSESSMENT: 'ASSESSMENT',
-  PROFILES_COLLECTION: 'PROFILES_COLLECTION',
-};
-
-export { CampaignTypes };
