+++ api/src/prescription/campaign/domain/validators/campaign-creation-validator.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
 import lodash from 'lodash';
 import { EntityValidationError } from '../../../../shared/domain/errors.js';
-import { CampaignTypes } from '../read-models/CampaignTypes.js';
+import { CampaignTypes } from '../../../shared/domain/constants.js';
 
 const { first } = lodash;
 const schema = Joi.object({
