+++ api/src/prescription/organization-learner/infrastructure/repositories/sup-organization-participant-repository.js
@@ -1,8 +1,10 @@
 import { knex } from '../../../../../db/knex-database-connection.js';
 import { fetchPage } from '../../../../../lib/infrastructure/utils/knex-utils.js';
 import { SupOrganizationParticipant } from '../../domain/read-models/SupOrganizationParticipant.js';
-import { CampaignTypes } from '../../../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
-import { CampaignParticipationStatuses } from '../../../../../src/prescription/shared/domain/constants.js';
+import {
+  CampaignTypes,
+  CampaignParticipationStatuses,
+} from '../../../../../src/prescription/shared/domain/constants.js';
 import { filterByFullName } from '../../../../../lib/infrastructure/utils/filter-utils.js';
 
 function _setFilters(qb, { search, studentNumber, groups, certificability } = {}) {
