+++ api/tests/integration/infrastructure/repositories/campaign-participation-overview-repository_test.js
@@ -4,8 +4,7 @@ const { campaignParticipationOverviewFactory } = databaseBuilder.factory;
 import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
 import * as campaignParticipationOverviewRepository from '../../../../lib/infrastructure/repositories/campaign-participation-overview-repository.js';
 import _ from 'lodash';
-import { CampaignParticipationStatuses } from '../../../../src/prescription/shared/domain/constants.js';
-import { CampaignTypes } from '../../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignParticipationStatuses, CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
 
 let userId;
