+++ api/tests/integration/infrastructure/repositories/participant-results-shared-repository_test.js
@@ -1,6 +1,6 @@
 import { expect, knex, databaseBuilder, mockLearningContent } from '../../../test-helper.js';
 import { participantResultsSharedRepository } from '../../../../lib/infrastructure/repositories/participant-results-shared-repository.js';
-import { CampaignTypes } from '../../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
 import { MAX_REACHABLE_PIX_BY_COMPETENCE } from '../../../../lib/domain/constants.js';
 
 describe('Integration | Repository | Campaign Participant Result Shared Repository', function () {
