+++ api/tests/prescription/campaign/unit/domain/validators/campaign-update-validator_test.js
@@ -1,6 +1,6 @@
 import { expect } from '../../../../../test-helper.js';
 import * as campaignUpdateValidator from '../../../../../../src/prescription/campaign/domain/validators/campaign-update-validator.js';
-import { CampaignTypes } from '../../../../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignTypes } from '../../../../../../src/prescription/shared/domain/constants.js';
 
 const MISSING_VALUE = null;
 const EMPTY_VALUE = '';
