+++ api/tests/prescription/organization-learner/integration/infrastructure/repositories/sco-organization-participant-repository_test.js
@@ -2,8 +2,10 @@ import _ from 'lodash';
 import { expect, databaseBuilder } from '../../../../../test-helper.js';
 import * as scoOrganizationParticipantRepository from '../../../../../../src/prescription/organization-learner/infrastructure/repositories/sco-organization-participant-repository.js';
 import { ScoOrganizationParticipant } from '../../../../../../src/prescription/organization-learner/domain/read-models/ScoOrganizationParticipant.js';
-import { CampaignTypes } from '../../../../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
-import { CampaignParticipationStatuses } from '../../../../../../src/prescription/shared/domain/constants.js';
+import {
+  CampaignParticipationStatuses,
+  CampaignTypes,
+} from '../../../../../../src/prescription/shared/domain/constants.js';
 
 describe('Integration | Infrastructure | Repository | sco-organization-participant-repository', function () {
   describe('#findPaginatedFilteredScoParticipants', function () {
