+++ api/tests/tooling/domain-builder/factory/build-campaign-report.js
@@ -1,4 +1,4 @@
-import { CampaignTypes } from '../../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
 import { CampaignReport } from '../../../../lib/domain/read-models/CampaignReport.js';
 
 const buildCampaignReport = function ({
