+++ api/tests/tooling/domain-builder/factory/build-organization-learner-participation.js
@@ -1,6 +1,5 @@
 import { OrganizationLearnerParticipation } from '../../../../lib/domain/read-models/OrganizationLearnerParticipation.js';
-import { CampaignTypes } from '../../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
-import { CampaignParticipationStatuses } from '../../../../src/prescription/shared/domain/constants.js';
+import { CampaignParticipationStatuses, CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
 
 const buildOrganizationLearnerParticipation = function ({
   id = '123',
