+++ api/tests/unit/domain/models/CampaignParticipation_test.js
@@ -1,7 +1,6 @@
 import { CampaignParticipation } from '../../../../lib/domain/models/CampaignParticipation.js';
-import { CampaignParticipationStatuses } from '../../../../src/prescription/shared/domain/constants.js';
 import { expect, domainBuilder, sinon, catchErr } from '../../../test-helper.js';
-import { CampaignTypes } from '../../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
+import { CampaignTypes, CampaignParticipationStatuses } from '../../../../src/prescription/shared/domain/constants.js';
 import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
 
 import {
