+++ api/tests/unit/domain/read-models/OrganizationLearnerActivity_test.js
@@ -1,7 +1,6 @@
 import { expect, domainBuilder } from '../../../test-helper.js';
 import { OrganizationLearnerActivity } from '../../../../lib/domain/read-models/OrganizationLearnerActivity.js';
-import { CampaignTypes } from '../../../../src/prescription/campaign/domain/read-models/CampaignTypes.js';
-import { CampaignParticipationStatuses } from '../../../../src/prescription/shared/domain/constants.js';
+import { CampaignParticipationStatuses, CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
 
 describe('Unit | Domain | Read-Models | OrganizationLearnerActivity', function () {
   it('computes the activity statistics for every status of every campaign types', function () {
